remove unneed interface method able to set thread pool size, gcd queue priority
git-svn-id: https://svn.ece.auckland.ac.nz/svn/se751-2012-g2/trunk@40 faf831e5-a2bb-0410-845a-b99f761cfef8

+++ ImageFilter/ImageFilter/BlurFilter.h
@@ -25,8 +25,8 @@ protected:
         const int RANGE = 10;
         for (int i = -RANGE; i < RANGE; i++) {
             for (int j = -RANGE; j < RANGE; j++) {
-                int x1 = x + i;
-                int y1 = y + j;
+                long x1 = x + i;
+                long y1 = y + j;
                 if (x1 < 0 || x1 >= w)
                     continue;
                 if (y1 < 0 || y1 >= h)
